Add prod compose file and update some stuff for tickets

+++ src/main/kotlin/beer/supportly/backend/dto/AgentTicketStatisticsDto.kt
@@ -13,6 +13,6 @@ package beer.supportly.backend.dto
 data class AgentTicketStatisticsDto(
     val globalTicketsOpen: Int,
     val yourTicketsOpen: Int,
-    val customerAccounts: Int,
+    val customerAccounts: Long,
     val averageTimePerTicket: Long
 )
